re-work mock for metata provider
