+++ tests/_support/font-awesome-phpunit-util.php
@@ -12,7 +12,7 @@ namespace FortAwesome;
 
 use \ReflectionException, \ReflectionProperty, \Exception;
 
-require_once FONTAWESOME_DIR_PATH . 'tests/_support/class-mock-fontawesome-releases.php';
+require_once FONTAWESOME_DIR_PATH . 'tests/_support/class-mock-fontawesome-metadata-provider.php';
 require_once FONTAWESOME_DIR_PATH . 'includes/class-fontawesome.php';
 require_once FONTAWESOME_DIR_PATH . 'includes/class-fontawesome-release-provider.php';
